fix: update context after sync if needed
A pull first update the peer's clock with the external context.
A sync (pull) may produce new operation.
This updates the clock.
The is_merging boolean block any synchronisation from the clock
to the context.
At the end of the pull we have to synchronise the clock and
the context if at least one operation was performed.